+++ SDLFramework/SDLFramework/Vector2D.h
@@ -8,7 +8,7 @@ public:
 	void sety(float y);
 	float getX() const;
 	float getY() const;
-	Vector2D multiply(Vector2D vector, int waarde);
+	Vector2D multiply(Vector2D vector, float waarde);
 	Vector2D Add(Vector2D vector, Vector2D vector2);
 	Vector2D sub(Vector2D vector, Vector2D vector2);
 	void div(float value);
